Added Razor bindings,  fixed the example app
The example app was using Escape.Uri instead of Filter.AsURL and also
over-escaped a lot of contexts due to Razor's autoescaping. The new
Razor bindings take care of disabling most of the auto-escaping (for
those contexts that could want it), but the escaping inside attributes
cannot be removed, so the examples needed to be changed.